refactor(renv): promote RuntimeEnvironment to pkg/renv (SRD-012 M2)
Move the per-execution RuntimeEnvironment interface from internal/renv to the
existing public pkg/renv (alongside EngineRuntime), and retire internal/renv.
The interface now embeds the public service.DataReader (the read half, SRD-011)
in place of its four explicit read methods, plus data.Source and the
write/identity/services tail; it imports only public packages
(service/data/eventproc/interactor), no cycle.

internal/instance.execEnv keeps implementing it (its read methods cover the
embedded DataReader; the rest comes via the embedded *Instance). Re-point the
9 importers (internal/exec, execenv, the 7 model executors) + 2 test files —
an import-path swap, refs unchanged (same package name renv). mockery + the
package doc updated.

pkg/model now imports no internal/renv (still internal/exec + internal/scope
until M3). Behaviour-preserving. Implements SRD-012 v.1 FR-2.
