+++ pkg/renv/engineruntime.go
@@ -1,12 +1,9 @@
-// Package renv defines EngineRuntime — the engine/server-level set of resolved
-// extensions (the wired services) that the Thresher owns and shares with the
-// things that run BPMN: instances (for node executors) and the EventHub (for
-// event waiters). It imports only the extension-interface packages, never the
-// execution machinery, so it can be referenced from internal/eventproc and
-// internal/instance without an import cycle (ADR-002 §4.3).
-//
-// The instance-level RuntimeEnvironment (which embeds this) stays in
-// internal/renv because it also exposes internal-only state.
+// Package renv defines the public runtime-environment contracts: EngineRuntime
+// — the engine/server-level set of resolved extensions (the wired services) the
+// Thresher owns and shares with the things that run BPMN — and the per-execution
+// RuntimeEnvironment a node executes against (ADR-012 §2.3). Both expose only
+// public types, so pkg/model implements/consumes them without importing
+// internal; the implementations live in internal/instance (ADR-002 §4.3).
 package renv
 
 import (
